refactor to reduce cognitive complexity
